ISIS-2727: moves ThrowingRunnable from commons to applib
