+++ api/applib/src/main/java/org/apache/isis/applib/services/sudo/SudoService.java
@@ -24,7 +24,7 @@ import java.util.function.UnaryOperator;
 import org.apache.isis.applib.services.iactnlayer.InteractionContext;
 import org.apache.isis.applib.services.user.RoleMemento;
 import org.apache.isis.applib.services.user.UserService;
-import org.apache.isis.commons.functional.ThrowingRunnable;
+import org.apache.isis.applib.services.iactnlayer.ThrowingRunnable;
 
 import lombok.NonNull;
