+++ commons/src/main/java/org/apache/isis/applib/services/iactnlayer/ThrowingRunnable.java
@@ -16,9 +16,11 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.apache.isis.commons.functional;
+package org.apache.isis.applib.services.iactnlayer;
 
 import java.util.concurrent.Callable;
+
+import org.apache.isis.commons.functional.Result;
 
 import lombok.NonNull;
